Move annotated text config components to Annotated/common
