+++ src/components/Text/Annotated/common/group/DefaultSegmentGroup.tsx
@@ -1,12 +1,12 @@
 import {
   projectConfigSelector,
   useProjectStore,
-} from "../../../../stores/project.ts";
-import { GroupProps } from "../core";
-import { isProjectAnnotation } from "../utils/isProjectAnnotation.ts";
+} from "../../../../../stores/project.ts";
+import { GroupProps } from "../../core";
+import { isProjectAnnotation } from "../../utils/isProjectAnnotation.ts";
 import { AnnotationLink } from "./AnnotationLink.tsx";
-import { useTextStore } from "../../../../stores/text/text-store.ts";
-import { orThrow } from "../../../../utils/orThrow.tsx";
+import { useTextStore } from "../../../../../stores/text/text-store.ts";
+import { orThrow } from "../../../../../utils/orThrow.tsx";
 
 export function DefaultSegmentGroup(props: GroupProps) {
   const projectConfig = useProjectStore(projectConfigSelector);
